modificado ft_substr

+++ libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: tmarina- <tmarina-@student.42urduliz.co    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/29 16:22:48 by tmarina-          #+#    #+#             */
-/*   Updated: 2022/12/26 17:57:55 by tmarina-         ###   ########.fr       */
+/*   Updated: 2022/12/29 12:31:36 by tmarina-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdio.h>
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	int	a;
+	size_t	a;
 
 	a = 0;
 	while (str[a])
